backend/pyproject.toml
[tool.poetry]
name = "autoflow-backend"
version = "0.1.0"
description = "AutoFlow Backend API and Execution Engine"
authors = ["AutoFlow Team <team@autoflow.dev>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = ">=0.100.0"
uvicorn = ">=0.23.0"
pydantic = ">=2.0.0"
pydantic-settings = ">=2.0.0"
sqlalchemy = ">=2.0.0"
alembic = ">=1.11.0"
python-dotenv = ">=1.0.0"
pyyaml = ">=6.0.0"
httpx = ">=0.24.0"
redis = ">=5.0.0"
celery = ">=5.3.0"
playwright = ">=1.36.0"
pyautogui = ">=0.9.54"
loguru = ">=0.7.0"
pymysql = ">=1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
black = ">=23.7.0"
isort = ">=5.12.0"
mypy = ">=1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
strict = true
